ARROW-16449: [Java] Prevent InaccessibleObjectException on Java 18
This change updates the implementation of MemoryUtil to not fail when
the JVM is not started with "--add-opens=java.base/java.nio=ALL-UNNAMED".

Previously, the static initializer of the MemoryUtil class would try to
make the "address" field of java.nio.Buffer accessible. However, since
sun.misc.Unsafe is used to access the field, the field does not need to
be accessible.

See Netty's approach as a reference:
https://github.com/netty/netty/commit/6ac5f35077925ebe2d76e9e2e1eab51a8b5eb35b